Added acceptance test for `ListIndex`
Fixed some type errors, which were preventing the tests to pass.

Moved the exising unit test for `ListIndex` and turned into an acceptance test. We're only using the API partially for `useBrowseSite()`, so I was able to reuse the existing mocking tools for it. The other requests are using `fetch` directly, which means we have to mock the request to it and respond with mocking data. I added mockiing data for inbox and following to the existing structure of mock responses.

This is only the first bit of tests, but it should give an idea on how we can proceed.